simplify: abort checks, deriveGatewayName extraction, comment trimming, async persist
- Replace ad-hoc AbortError checks with `isAbortError` in connect-upstream-dialog and install-dialog
- Move `deriveGatewayName` (and its regex constants) from install-dialog to gateway-name.ts
- Trim 8-line upstream_oauth.rs comment block to one line; reuse `now` instead of re-computing `revoke_now`
- Trim 7-line sqlite.rs doc comment to one sentence
- Remove narrative what-comments from mcpregistry/dispatch.rs
- Spawn persist_config detached in manager.rs so OAuth redirect completes without waiting for disk write
