+++ apps/gateway-admin/lib/utils/gateway-name.ts
@@ -9,6 +9,8 @@
  *   - max 64 characters
  */
 const GATEWAY_NAME_PATTERN = /^[a-zA-Z0-9][a-zA-Z0-9_-]{0,63}$/
+const BIDI_STRIP_RE = /[\u200B-\u200F\u202A-\u202E\u2060-\u2064\uFEFF]/g
+const INVALID_CHARS_RE = /[^a-zA-Z0-9_-]/g
 
 /**
  * Validate a gateway name.
@@ -19,4 +21,12 @@ export function validateGatewayName(value: string): string | null {
   if (!GATEWAY_NAME_PATTERN.test(value))
     return 'Must start with a letter or digit and contain only letters, digits, underscores, or hyphens (max 64 chars)'
   return null
+}
+
+export function deriveGatewayName(serverName: string): string {
+  const segment = serverName.split('/').at(-1) ?? serverName
+  return segment
+    .normalize('NFC')
+    .replace(BIDI_STRIP_RE, '')
+    .replace(INVALID_CHARS_RE, '')
 }
